Added Stripe warnings to startview
